% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleD3Bar.R
\name{simpleD3Bar}
\alias{simpleD3Bar}
\title{D3 bar chart}
\usage{
simpleD3Bar(data, colors = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{data to be sent to JS}

\item{colors}{Colors to be sent to JS}

\item{width}{Width if specified}

\item{height}{Height if specified}

\item{elementId}{Id if specified}
}
\description{
Just a simple bar chart.
}
