% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleD3Legend.R
\name{simpleD3Legend}
\alias{simpleD3Legend}
\title{D3 legend}
\usage{
simpleD3Legend(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{data to be sent to JS}

\item{width}{Width if specified}

\item{height}{Height if specified}

\item{elementId}{Id if specified}
}
\description{
Just a simple legendT.
}
