% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleD3Line.R
\name{simpleD3Line-shiny}
\alias{simpleD3Line-shiny}
\alias{simpleD3LineOutput}
\title{Shiny bindings for simpled3}
\usage{
simpleD3LineOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a simpled3}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using simpled3 within Shiny
applications and interactive Rmd documents.
}
