% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleD3Line.R
\name{simpleD3Line}
\alias{simpleD3Line}
\title{D3 line}
\usage{
simpleD3Line(data, colors = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{data to be sent to JS}

\item{colors}{Colors to be sent to JS}

\item{width}{Width if specified}

\item{height}{Height if specified}

\item{elementId}{Id if specified}
}
\description{
Just a simple legend.
}
